+++ scripts/common/constants/VolforceRanks.lua
@@ -1,62 +1,52 @@
 ---@type { color: Color, name: string, step: number, start: integer }[]
 local VolforceRanks = {
   {
-    color = { 150, 100, 195 },
-    name = "IMPERIAL",
+    name = "[ IMPERIAL ]",
     start = 20,
     step = 1.0,
   },
   {
-    color = { 195, 50, 50 },
-    name = "CRIMSON",
+    name = "[ CRIMSON ]",
     start = 19,
     step = 0.25,
   },
   {
-    color = { 245, 230, 150 },
-    name = "ELDORA",
+    name = "[ ELDORA ]",
     start = 18,
     step = 0.25,
   },
   {
-    color = { 220, 225, 230 },
-    name = "ARGENTO",
+    name = "[ ARGENTO ]",
     start = 17,
     step = 0.25,
   },
   {
-    color = { 230, 150, 175 },
-    name = "CORAL",
+    name = "[ CORAL ]",
     start = 16,
     step = 0.25,
   },
   {
-    color = { 210, 60, 60 },
-    name = "SCARLET",
+    name = "[ SCARLET ]",
     start = 15,
     step = 0.25,
   },
   {
-    color = { 55, 175, 175 },
-    name = "CYAN",
+    name = "[ CYAN ]",
     start = 14,
     step = 0.25,
   },
   {
-    color = { 240, 170, 30 },
-    name = "DANDELION",
+    name = "[ DANDELION ]",
     start = 12,
     step = 0.5,
   },
   {
-    color = { 60, 95, 175 },
-    name = "COBALT",
+    name = "[ COBALT ]",
     start = 10,
     step = 0.5,
   },
   {
-    color = { 180, 130, 80 },
-    name = "SIENNA",
+    name = "[ SIENNA ]",
     start = 0,
     step = 2.5,
   },
@@ -79,7 +69,7 @@ function VolforceRanks:get(volforce)
     end
   end
 
-  return "RANK"
+  return "[ RANK ]"
 end
 
 return VolforceRanks